pyproject.toml
[project]
name = "comic-translate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.25.6",
    "azure-ai-vision-imageanalysis>=1.0.0b1",
    "celery>=5.4.0",
    "dayu-path>=0.5.2",
    "deep-translator>=1.11.4",
    "deepl>=1.16.1",
    "easyocr>=1.7.1",
    "ebooklib>=0.18",
    "fastapi>=0.115.8",
    "fugashi>=1.4.0",
    "google-generativeai>=0.5.2",
    "img2pdf>=0.5.1",
    "jaconv>=0.3.4",
    "largestinteriorrectangle>=0.2.0",
    "loguru>=0.7.2",
    "msgpack>=1.1.0",
    "openai>=1.8.0",
    "opencv-python>=4.11.0.86",
    "paddleocr>=2.8.1",
    "paddlepaddle>=2.6.1",
    "py7zr>=0.20.8",
    "pymupdf>=1.24.9",
    "pyperclip>=1.9.0",
    "pyside6>=6.8.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "rarfile>=4.1",
    "requests>=2.31.0",
    "stanza>=1.7.0",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "transformers>=4.36.2",
    "ultralytics<=8.3.40",
    "unidic-lite>=1.0.8",
    "uvicorn>=0.34.0",
    "wget>=3.2",
]
